Modify Dataset API structure
