+++ src/otx/v2/adapters/torch/mmengine/mmpretrain/__init__.py
@@ -16,7 +16,7 @@ try:
 
     register_all_modules(init_default_scope=True)
 
-    from .dataset import Dataset
+    from .dataset import MMPretrainDataset as Dataset
     from .engine import MMPTEngine as Engine
     from .model import get_model, list_models
